chore: change file structure
to accomodate more match funcs
