+++ snaps/utils_test.go
@@ -1,48 +1,10 @@
 package snaps
 
 import (
-	"sync"
 	"testing"
 
 	"github.com/gkampitakis/go-snaps/snaps/internal/test"
 )
-
-func TestUtils(t *testing.T) {
-	t.Run("should create a string from multiple params", func(t *testing.T) {
-		expected := "test\nint(5)\nmap[string]string{\"test\":\"test\"}\n"
-		received := takeSnapshot([]interface{}{"test", 5, map[string]string{"test": "test"}})
-
-		test.Equal(t, expected, received)
-	})
-
-	t.Run("getTestID should increment id on each call [concurrent safe]", func(t *testing.T) {
-		wg := sync.WaitGroup{}
-		registry := newRegistry()
-
-		for i := 0; i < 5; i++ {
-			wg.Add(1)
-
-			go func() {
-				registry.getTestID("test", "/file")
-				wg.Done()
-			}()
-		}
-
-		wg.Wait()
-
-		test.Equal(t, "[test - 6]", registry.getTestID("test", "/file"))
-		test.Equal(t, "[test-v2 - 1]", registry.getTestID("test-v2", "/file"))
-	})
-
-	t.Run("dynamicTestIDRegexp", func(t *testing.T) {
-		t.Run("should escape regular expressions from testID", func(t *testing.T) {
-			escapedTestIDRegexp := dynamicTestIDRegexp(`^\s+$-test`)
-			expectedRegexp := `(?m)(?:\^\\s\+\$-test[\s\S])(.*[\s\S]*?)(?:^---$)`
-
-			test.Equal(t, expectedRegexp, escapedTestIDRegexp.String())
-		})
-	})
-}
 
 func TestBaseCallerNested(t *testing.T) {
 	file := baseCaller(0)
@@ -80,5 +42,35 @@ func TestBaseCaller(t *testing.T) {
 
 	t.Run("should return function's name", func(t *testing.T) {
 		TestBaseCallerNested(t)
+	})
+}
+
+func TestEnv(t *testing.T) {
+	t.Run("should return true if env var is 'true'", func(t *testing.T) {
+		test.SetEnv(t, "MOCK_ENV", "true")
+
+		res := getEnvBool("MOCK_ENV", false)
+
+		if !res {
+			t.Error("getEnvBool should return true")
+		}
+	})
+
+	t.Run("should return false", func(t *testing.T) {
+		test.SetEnv(t, "MOCK_ENV", "")
+
+		res := getEnvBool("MOCK_ENV", true)
+
+		if res {
+			t.Error("getEnvBool should return false")
+		}
+	})
+
+	t.Run("should return fallback value for non existing env var", func(t *testing.T) {
+		res := getEnvBool("MISSING_ENV", true)
+
+		if !res {
+			t.Error("getEnvBool should return false")
+		}
 	})
 }
